requirements.txt
# Core dependencies
fastapi>=0.117.1
uvicorn>=0.34.2
httpx>=0.28.1
python-dotenv>=1.1.0
jinja2>=3.1.0

# OpenAI and AI
openai>=1.57.0

# GitHub integration
pygithub>=2.5.0

# Data handling
pydantic>=2.11.4

# A2A Framework (if using the full framework)
a2a-sdk[http-server]>=0.3.0

# Additional utilities
click>=8.1.8
requests>=2.31.0
